requirements.txt
pandas==1.3.5
numpy==1.21.5
torch==1.10.0
torchaudio=0.10.0
torchvision=0.11.1
torchsummary==1.5.1
torchtext==0.11.0
transformers==4.17.0
sklearn-pandas==1.8.0
scikit-learn==1.0.2